applied changes from PR comments

+++ database/migrations/20250720183611-create-users-table.js
@@ -3,7 +3,7 @@
 module.exports = {
   up: async (queryInterface, Sequelize) => {
     await queryInterface.createTable('Users', {
-      user_id: { type: Sequelize.INTEGER, autoIncrement: true, primaryKey: true },
+      user_id: { type: Sequelize.BIGINT, autoIncrement: true, primaryKey: true },
       email: { type: Sequelize.STRING, allowNull: false },
       password: { type: Sequelize.STRING, allowNull: false },
     });
